move mockNeo4jMCPServer into the test file it's used in
